Remove old fitswcs file which isn't needed anymore and update tests
