+++ astropy/visualization/wcsaxes/tests/test_wcsapi.py
@@ -4,12 +4,144 @@ import pytest
 import numpy as np
 
 import matplotlib.pyplot as plt
+from matplotlib.transforms import Affine2D, IdentityTransform
 
+from astropy import units as u
 from astropy.wcs.wcsapi import BaseLowLevelWCS
 from astropy.coordinates import SkyCoord
 from astropy.time import Time
 from astropy.units import Quantity
 from astropy.tests.image_tests import IMAGE_REFERENCE_DIR
+from astropy.wcs import WCS
+from astropy.visualization.wcsaxes.frame import RectangularFrame
+from astropy.visualization.wcsaxes.wcsapi import (WCSWorld2PixelTransform,
+                                                  transform_coord_meta_from_wcs)
+
+WCS2D = WCS(naxis=2)
+WCS2D.wcs.ctype = ['x', 'y']
+WCS2D.wcs.cunit = ['km', 'km']
+WCS2D.wcs.crpix = [614.5, 856.5]
+WCS2D.wcs.cdelt = [6.25, 6.25]
+WCS2D.wcs.crval = [0., 0.]
+
+WCS3D = WCS(naxis=3)
+WCS3D.wcs.ctype = ['x', 'y', 'z']
+WCS3D.wcs.cunit = ['km', 'km', 'km']
+WCS3D.wcs.crpix = [614.5, 856.5, 333]
+WCS3D.wcs.cdelt = [6.25, 6.25, 23]
+WCS3D.wcs.crval = [0., 0., 1.]
+
+
+def test_shorthand_inversion():
+    """
+    Test that the Matplotlib subtraction shorthand for composing and inverting
+    transformations works.
+    """
+    w1 = WCS(naxis=2)
+    w1.wcs.ctype = ['RA---TAN', 'DEC--TAN']
+    w1.wcs.crpix = [256.0, 256.0]
+    w1.wcs.cdelt = [-0.05, 0.05]
+    w1.wcs.crval = [120.0, -19.0]
+
+    w2 = WCS(naxis=2)
+    w2.wcs.ctype = ['RA---SIN', 'DEC--SIN']
+    w2.wcs.crpix = [256.0, 256.0]
+    w2.wcs.cdelt = [-0.05, 0.05]
+    w2.wcs.crval = [235.0, +23.7]
+
+    t1 = WCSWorld2PixelTransform(w1)
+    t2 = WCSWorld2PixelTransform(w2)
+
+    assert t1 - t2 == t1 + t2.inverted()
+    assert t1 - t2 != t2.inverted() + t1
+    assert t1 - t1 == IdentityTransform()
+
+
+# We add Affine2D to catch the fact that in Matplotlib, having a Composite
+# transform can end up in more strict requirements for the dimensionality.
+
+
+def test_2d():
+
+    world = np.ones((10, 2))
+
+    w1 = WCSWorld2PixelTransform(WCS2D) + Affine2D()
+    pixel = w1.transform(world)
+    world_2 = w1.inverted().transform(pixel)
+
+    np.testing.assert_allclose(world, world_2)
+
+
+def test_3d():
+
+    world = np.ones((10, 2))
+
+    w1 = WCSWorld2PixelTransform(WCS3D[:, 0, :]) + Affine2D()
+    pixel = w1.transform(world)
+    world_2 = w1.inverted().transform(pixel)
+
+    np.testing.assert_allclose(world[:, 0], world_2[:, 0])
+    np.testing.assert_allclose(world[:, 1], world_2[:, 1])
+
+
+CTYPE_CASES = [(' LON-TAN', ('longitude', None, None)),
+               (' LAT-TAN', ('latitude', None, None)),
+               ('HPLN-TAN', ('longitude', u.arcsec, 180.)),
+               ('HPLT-TAN', ('latitude', u.arcsec, None)),
+               ('RA---TAN', ('longitude', u.hourangle, None)),
+               ('DEC--TAN', ('latitude', None, None)),
+               ('spam', ('scalar', None, None))]
+
+
+def test_coord_type_from_ctype():
+
+    wcs = WCS(naxis=2)
+    wcs.wcs.ctype = ['GLON-TAN', 'GLAT-TAN']
+    wcs.wcs.crpix = [256.0] * 2
+    wcs.wcs.cdelt = [-0.05] * 2
+    wcs.wcs.crval = [50.0] * 2
+
+    _, coord_meta = transform_coord_meta_from_wcs(wcs, RectangularFrame)
+
+    assert coord_meta['type'] == ['longitude', 'latitude']
+    assert coord_meta['format_unit'] == [u.deg, u.deg]
+    assert coord_meta['wrap'] == [None, None]
+
+    wcs = WCS(naxis=2)
+    wcs.wcs.ctype = ['HPLN-TAN', 'HPLT-TAN']
+    wcs.wcs.crpix = [256.0] * 2
+    wcs.wcs.cdelt = [-0.05] * 2
+    wcs.wcs.crval = [50.0] * 2
+
+    _, coord_meta = transform_coord_meta_from_wcs(wcs, RectangularFrame)
+
+    assert coord_meta['type'] == ['longitude', 'latitude']
+    assert coord_meta['format_unit'] == [u.arcsec, u.arcsec]
+    assert coord_meta['wrap'] == [180., None]
+
+    wcs = WCS(naxis=2)
+    wcs.wcs.ctype = ['RA---TAN', 'DEC--TAN']
+    wcs.wcs.crpix = [256.0] * 2
+    wcs.wcs.cdelt = [-0.05] * 2
+    wcs.wcs.crval = [50.0] * 2
+
+    _, coord_meta = transform_coord_meta_from_wcs(wcs, RectangularFrame)
+
+    assert coord_meta['type'] == ['longitude', 'latitude']
+    assert coord_meta['format_unit'] == [u.hourangle, u.deg]
+    assert coord_meta['wrap'] == [None, None]
+
+    wcs = WCS(naxis=2)
+    wcs.wcs.ctype = ['spam', 'spam']
+    wcs.wcs.crpix = [256.0] * 2
+    wcs.wcs.cdelt = [-0.05] * 2
+    wcs.wcs.crval = [50.0] * 2
+
+    _, coord_meta = transform_coord_meta_from_wcs(wcs, RectangularFrame)
+
+    assert coord_meta['type'] == ['scalar', 'scalar']
+    assert coord_meta['format_unit'] == [u.one, u.one]
+    assert coord_meta['wrap'] == [None, None]
 
 
 class LowLevelWCS5D(BaseLowLevelWCS):
